1.3.Security & the User Class{Creating the User Class with make:user}
Creating the User Class with make:user
let's try it!

php bin/console make:user

Call the class User. Second question:

Do you want to store user data in the database
For most apps, this is an easy yes... because most apps store user data in a local database table. But, what if your user data is stored on some other server, like an LDAP server or a single sign-on server? Well, even in those cases, if you want to store any extra information about your users in a local database table, you should still answer yes. Answer "no" only if you don't need to store any user information to your database.
So, "yes" for us! Next: choose one property on your user that will be its unique display name. This can be anything - it's usually an email or username. We'll talk about how it's used later. Choose email.
And, the last question: is our app responsible for checking the user's password? In some apps - like a pure API with only token authentication, users might not even have a password. And even if your users will be able to login with a password, only answer yes if this app will be responsible for directly checking the user's password. If you actually send the password to a third-party server and it checks if it's valid, choose no.
Remember when I mentioned how complex & different modern authentication systems can be? That's why this command exists: to help walk us through exactly want we need.
I'm going to choose "No" for now. We will add a password later, but we'll keep things extra simple to start.
And... we're done! Awesome! This created a User entity, a Doctrine UserRepository for it, and updated the security.yaml file.