3.1.Customizing the User Entity
The really neat thing about Symfony's security system is that it doesn't care at all about what your User class looks like. As long as it implements UserInterface, so, as long as it has these methods, you can do anything you want with it. Heck, it doesn't even need to be an entity!

Adding more Fields to User
For example, we already have an email field, but I also want to be able to store the first name for each user. Cool: we can just add that field! Find your terminal and run:

php bin/console make:entity
Update the User class and add firstName as a string, length 255 - or shorter if you want - and not nullable. Done!
Check out the User class! Yep, there's the new firstName property and... at the bottom, the getter and setter methods.
Awesome!

Setting Doctrine's server_version
I think we're ready to make the migration. But! A word of warning. Check out the roles field on top.
It's an array and its Doctrine type is json. This is really cool. Newer databases - like PostgreSQL and MySQL 5.7 - have a native JSON column type that allows you to store an array of data.
But, if you're using MySQL 5.6 or lower, this column type does not exist. And actually, that's not a problem! In that case, Doctrine is smart enough to use a normal text field, json_encode() your array when saving, and json_decode() it automatically when we query. So, no matter what database you use, you can use this json Doctrine column type.
But, here's the catch. Open config/packages/doctrine.yaml. One of the keys here is server_version, which is set to 5.7 by default
This tells Doctrine that when it interacts with the database, it should expect that our database has all the features supported by MySQL 5.7, including that native JSON column type. If your computer, or more importantly, if your production database is using MySQL 5.6, then you'll get a huge error when Doctrine tries to make queries using the native MySQL JSON column type.
If you're in this situation, just set this back to 5.6. Doctrine will then create a normal text column for the JSON field.