4.2.The Login Form{Filling in the Security Logic & Login Form}
Empty all of the existing code from the template. Then, change the title to Login! and, for now, just add an h1 with "Login to the SpaceBar! ".

Filling in the Security Logic & Login Form
Try it again: perfect! Well, not perfect - it looks terrible... and there's no login form yet. To fix that part, Google for "Symfony login form" to find a page on the Symfony docs that talks all about this. We're coming here so that we can steal some code!
Scroll down a bit until you see a login() method that has some logic in it. Copy the body, move back to our controller, and paste! This needs an AuthenticationUtils class as an argument. Add it: AuthenticationUtils . Then, these two new variables are passed into Twig. Copy them, and also paste it.
In a few minutes, we're going to talk about where these two variables are set. They both deal with authentication.
But first, go back to the docs and find the login form. Copy this, move over and paste it into our body.
Notice: there is nothing special about this form: it has a username field, a password field and a submit button. And, we're going to customize it, so don't look too closely yet.
Move back to your browser to check things out. Bah!

  Unable to generate a URL for the named route "login"

This comes from login.html.twig. Of course! The template we copied is pointing to a route called login, but our route is called app_login