4.3.The Login Form{Filling in the Security Logic & Login Form; Including the login.css File}
The template we copied is pointing to a route called login, but our route is called app_login.
Actually, just remove the action= entirely.
If a form doesn't have an action attribute, it will submit right back to the same URL - /login - which is what I want anyways.
Refresh again. Perfect! Well, it still looks awful. Oof. To fix that, I'm going to replace the HTML form with some markup that looks nice in Bootstrap 4 - you can copy this from the code block on this page.

Including the login.css File
Before we look at this new code, try it! Refresh! Still ugly! Dang! Oh yea, that's because we need to include a new CSS file for this markup.
If you downloaded the course code, you should have a tutorial/ directory with two CSS files inside. Copy login.css, find your public/ directory and paste the file into public/css.
So far in this series, we are not using Webpack Encore, which is an awesome tool for professionally combining and loading CSS and JS files. Instead, we're just putting CSS files into the public/ directory and pointing to them directly. If you want to learn more about Encore, go check out our Webpack Encore tutorial.
Anyways, we need to add a link tag for this new CSS file... but I only want to include it on this page, not on every page - we just don't need the CSS on every page. Look at base.html.twig. We're including three CSS files in the base layout. Ah, and they all live inside a block called stylesheets.
We basically want to add a fourth link tag right below these... but only on the login page. To do that, in login.html.twig, add block stylesheets and endblock. This will override that block completely... which is actually not exactly what we want. Nope, we want to add to that block. To do that print parent()
This will print the content of the parent block - the 3 link tags - and then we can add the new link tag below: link, with href= and login.css. PhpStorm helps fill in the asset() function.
Now it should look good. Try it. Boom! Oh, but we don't need that h1 tag anymore.