4.4.The Login Form{The Fields of the Login Form}
The Fields of the Login Form
So even though this looks much better, it's still just a very boring HTML form. It has an email field and a password field... though, we won't add the password-checking logic until later. It also has a "remember me" checkbox that we'll learn how to activate.
The point is: you can make your login form look however you want. The only special part is this error variable, which, when we're done, will be the authentication error if the user just entered a bad email or password. I'll plan ahead and add a Bootstrap class for this.