4.5.The Login Form{Adding a Link to the Login Page}
Adding a Link to the Login Page
Ok. Login form is done! But... we probably need a link to this page. In the upper right corner, we have a cute user dropdown... which is totally hardcoded with fake data. Go back to base.html.twig and scroll down to find this. There it is! For now, let's comment-out that drop-down. We'll re-add it later when we have real data. Then, copy a link from above, paste ithere and change it to Login with a link to app_login.
Try it - refresh! We got it! HTML login form, check! We are now ready to fill in the logic of what happens when we submit the form. We'll do that in something called an "authenticator".