5.3.Firewalls & Authenticator{Creating the Authentication with make:auth}
This class is awesome: it basically has a method for each step of the authentication process. Before we walk through each one, because this authenticator will be for a login form, there's a different base class that allows us to... well... do less work!
Instead of extends AbstractGuardAuthenticator use extends AbtractFormLoginAuthenticator. I'll remove the old use statement.
Thanks to this, we no longer need onAuthenticationFailure(), start() or supportsRememberMe(): they're all handled for us.
But don't worry, when we create an API token authenticator later, we will learn about these methods. We do now need one new method. Go to the "Code"->"Generate" menu, or Command+N on a Mac, and select "Implement Methods" to generate getLoginUrl()