5.4.Firewalls & Authenticator{Activating the Authenticator in security.yaml}
Activating the Authenticator in security.yaml
Perfect! Unlike a lot of features in Symfony, this authenticator won't be activated automatically. To tell Symfony about it, go back to security.yaml. Under the main firewall, add a new guard key, a new authenticators key below that, and add one item in that array: App\Security\LoginFormAuthenticator.
The whole authenticator system comes from a part of the Security component called "Guard", hence the name. The important part is that, as soon as we add this, at the beginning of every request, Symfony will call the supports() method on our authenticator.
To prove it, add a die() statement with a message.
Then, move over and, refresh! Got it! And it doesn't matter what URL we go to: the supports() method is always called at the start of the request.
And now, we're in business! Let's fill in these methods and get our user logged in.