6.1.Login Form Authenticator{The supports() Method}
Now that we've added our authenticator under the authenticators key. Symfony calls its supports() method at the beginning of every request, which is why we see this little die statement.These authenticator classes are really cool because each method controls just one small part of the authentication process.

The supports() Method
The first method - supports() - is called on every request. Our job is simple: to return true if this request contains authentication info that this authenticator knows how to process. And if not, to return false.
In this case, when we submit the login form, it POSTs to /login. So, our authenticator should only try to authenticate the user in that exact situation. Return $request->attributes->get('_route') === 'app_login'
Let me... explain this. If you look in SecurityController, the name of our login route is app_login.
And, though you don't need to do it very often, if you want to find out the name of the currently-matched route, you can do that by reading this special _route key from the request attributes. In other words, this is checking to see if the URL is /login. We also only want our authenticator to try to login the user if this is a POST request. So, add && ->isMethod('POST').
Here's how this works: if we return false from supports(), nothing else happens. Symfony doesn't call any other methods on our authenticator, and the request continues on like normal to our controller, like nothing happened. It's not an authentication failure - it's just that nothing happens at all.
If we return true from supports(), well, that's when the fun starts. If we return true, Symfony will immediately call getCredentials().
To see if things are working, let's just dump($request->request->all()), then die().I know, that looks funny. Unrelated to security, if you want to read POST data off of the request, you use the $request->request property.
Anyways, let's try it! Go back to your browser and hit enter on the URL so that it makes a GET request to /login. Hello login page! Our supports() method just returned false. And so, the request continued anonymously, like normal.
Log in with one of our dummy users: spacebar1@example.com. The password doesn't matter. And... enter! Yes! This time, because this is a POST request to /login, supports() returns true! So, Symfony calls getCredentials() and our dump fires! As expected, we can see the email and password POST parameters, because the login form uses these names.