6.2.Login Form Authenticator{The Brand-New dd() Function}
The Brand-New dd() Function
Oh, and I want to show you a quick new Easter egg in Symfony 4.1, unrelated to security. Instead of dump() and die, use dd() and then remove the die.
Refresh! Same result. This is just a nice, silly shortcut: dd() is dump() and die. We'll use it... because... why not?