6.3.Login Form Authenticator{The getCredentials() Method}
The getCredentials() Method
Back to work! Our job in getCredentials() is simple: to read our authentication credentials off of the request and return them. In this case, we'll return the email and password. But, if this were an API token authenticator, we would return that token. We'll see that later.
Return an array with an email key set to $request->request->get('email') and password set to $request->request->get('password').
I'm just inventing these email and password keys for the new array: we can really return whatever we want from this method. Because, after we return from getCredentials(), Symfony will immediately call getUser() and pass this array back to us as the first $credentials argument. Let's see that in action: dd($credentials). Move back to your browser and, refresh! Coincidentally, it dumps the exact same thing as before. But, this time, it's coming from line 24 - our line in getUser().