6.4.Login Form Authenticator{The getUser() Method}
The getUser() Method
Let's keep going! Our job in getUser() is to use these  to return a User object, or null if the user isn't found. Because we're storing our users in the database, we need to query for the user via their email. And to do that, we need the UserRepository that was generated with our entity.
At the top of the class, add public function __construct() with a UserRepository $userRepository argument. I'll hit Alt+Enter and select "Initialize Fields" to add that property and set it.
Back down in getUser(), just return $this->userRepository->findOneBy() to query by email, set to $credentials['email'].
This will return our User object, or null. The cool thing is that if this returns null, the whole authentication process will stop, and the user will see an error. But if we return a User object, then Symfony immediately calls checkCredentials(), and passes it the same $credentials and the User object we just returned.
Inside, dd($user) so we can see if things are working. Refresh and... got it! That's our User object!