6.5.Login Form Authenticator{The checkCredentials() Method}
The checkCredentials() Method
Ok, final step: checkCredentials(). This is your opportunity to check to see if the user's password is correct, or any other last, security checks. Right now... well... we don't have a password, so, let's return true.
And actually, in many systems, simply returning true is perfect! For example, if you have an API token system, there's no password.
If you did return false, authentication would fail and the user would see an "Invalid Credentials" message. We'll see that soon.
But, when you return true... authentication is successful! Woo! To figure out what to do, now that the user is authenticated, Symfony calls onAuthenticationSuccess().
Put a dd() here that says "Success"
Move over and... refresh the POST! Yes! We hit it! At this point, we have fully filled in all the authentication logic. We used supports() to tell Symfony whether or not our authenticator should be used in this request, fetched credentials off of the request, used those to find the user, and returned true in checkCredentials() because we don't have a password.
Next, let's fill in these last two methods and finally see - for real - that our user is logged in. We'll also learn a bit more about what happens when authentication fails and how the error message is rendered.