7.2.Redirecting on Success & the User Provider{Redirecting on Success; Authentication & the Session: User Provider}
But, of course, we don't ever hardcode URLs in Symfony. Instead, we need to generate a URL to the route named app_homepage.
We know how to generate URLs in Twig - the path() function. But, how can we do it in PHP? The answer is... with Symfony's router service. To find out how to get it, run:

php bin/console debug:autowiring
Look for something related to routing... there it is! Actually, there are a few different router-related interfaces... but they're all different ways to get the same service. I usually use RouterInterface.
Back on top, add a second constructor argument: RouterInterface $router. I'll hit Alt+Enter and select "Initialize Fields" to create that property and set it.
Then, back down below, use $this->router->generate() to make a URL to app_homepage. Ok! We still have one empty method.
But, forget that! We're ready! Go back to your browser, and hit enter to show the login page again. Let's walk through the entire process. Use the same email, any password and... enter! It worked! How do I know? Check out the web debug toolbar! We are logged in as spacebar1@example.com!

Authentication & the Session: User Provider
This is even cooler than it looks. Think about it: we made a POST request to /login and became authenticated thanks to our authenticator. Then, we were redirected to the homepage... where our authenticator did nothing, because its supports() method returned false.
The only reason we're still logged in - even though our authenticator did nothing on this request - is that user authentication info is stored to the session. At the beginning of every request, that info is loaded from the session and we're logged in. Cool!
Look back at your security.yaml file. Remember this user provider thing that was setup for us?
This is a class that helps with the process of loading the user info from the session.
Honestly, it's a little bit confusing, but super important. Here's the deal: when you refresh the page, the User object is loaded from the session. But, we need to make sure that the object isn't out of date with the database. Think about it. Imagine we login at work. Then, we login at home and update our first name in the database. The next day, when we go back to work, we reload the page. Well... if we did nothing else, the User object we reloaded from the session for that browser would have our old first name. That would probably cause some weird issues.
So, that's the job of the user provider. When we refresh, the user provider takes the User object from the session and uses its id to query for a fresh User object. It all happens invisibly, which is great. But it is an important, background detail.
Next, I want to see what happens when we fail authentication. What does the user see? How are errors displayed? And how can we control them?