8.2.Authentication Errors{Filling in getLoginUrl()}
Filling in getLoginUrl()
If you dug a bit, you'd find out that, on failure, that authenticator class is calling getLoginUrl() and trying to redirect there. And, yea, that makes sense: if we fail login, the user should be redirected back to the login page. To make this actually work, all we need to do is fill in this method.
No problem: return $this->router->generate('app_login').
Ok, try it again: refresh and... perfect! Hey! You can even see an error message on top:

  Username could not be found.

We get that exact error because of where the authenticator fails: we failed to return a user from getUser().
In a little while, we'll learn how to customize this message because... probably saying "Email" could not be found would make more sense.