8.4.Authentication Errors{How Authentication Errors are Stored}
Anyways, go change checkCredential response back to true.
How Authentication Errors are Stored
What I really want to find out is: where are these errors coming from? In SecurityController, we're getting the error by calling some $authenticationUtils->getLastAuthenticationError() method:
28 lines  src/Controller/SecurityController.php
use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
class SecurityController extends AbstractController
{
    public function login(AuthenticationUtils $authenticationUtils)
    {
        // get the login error if there is one
        $error = $authenticationUtils->getLastAuthenticationError();
        return $this->render('security/login.html.twig', [
            'error'         => $error,
        ]);
    }
}
We're passing that into the template and rendering its messageKey property... with some translation magic we'll talk about soon too:
32 lines  templates/security/login.html.twig
{% block body %}
    <form class="form-signin" method="post">
        {% if error %}
            <div class="alert alert-danger">{{ error.messageKey|trans(error.messageData, 'security') }}</div>
        {% endif %}
    </form>
{% endblock %}
The point is: we magically fetch the "error" from... somewhere and render it. Let's demystify that. Go back to the top of your authenticator and hold command or control to click into AbstractFormLoginAuthenticator.
In reality, when authentication fails, this onAuthenticationFailure() method is called. It's a bit technical, but when authentication fails, internally, it's because something threw an AuthenticationException, which is passed to this method. And, ah: this method stores that exception onto a special key in the session! Then, back in the controller, the lastAuthenticationError() method is just a shortcut to read that key off of the session!
So, it's simple: our authenticator stores the error in the session and then we read the error from the session in our controller and render it:
28 lines  src/Controller/SecurityController.php
class SecurityController extends AbstractController
{
    public function login(AuthenticationUtils $authenticationUtils)
    {
        // get the login error if there is one
        $error = $authenticationUtils->getLastAuthenticationError();
    }
}
The last thing onAuthenticationFailure() does is call our getLoginUrl() method and redirect there.