8.5.Authentication Errors{Filling in the Last Email}
Filling in the Last Email
Go back to the login form and fail authentication again with a fake email. We see the error... but the email field is empty - that's not ideal. For convenience, it should pre-fill with the email I just entered.
Look at the controller again. Hmm: we are calling a getLastUsername() method and passing that into the template.
Oh, but I forgot to render it! Add value= and print last_username.
But... we're not quite done. Unlike the error message, the last user name is not automatically stored to the session. This is something that we need to do inside of our LoginFormAuthenticator. But, it's super easy. Inside getCredentials(), instead of returning, add $credentials =
Now, set the email onto the session with $request->getSession()->set(). Use a special key: Security - the one from the Security component - ::LAST_USERNAME and set this to $credentials['email'].
Then, at the bottom, return $credentials.
Try it! Go back, login with that same email address and... nice! Both the error and the last email are read from the session and displayed.
Next: let's learn how to customize these error messages. And, we really need a way to logout.