9.2.Customizing Errors & Logout{Logging Out}
Logging Out
Hey! Our login authentication system is... done! And... not that I want to rush our moment of victory - we did it! - but now that our friendly alien users can log in... they'll probably need a way to log out. They're just never satisfied...
Right now, I'm still logged in as spacebar1@example.com. Let's close a few files. Then, open SecurityController. Step 1 to creating a logout system is to create the route. Add public function logout().
Above this, use the normal @route("/logout") with the name app_logout.
And this is where things get interesting... We do need to create this route... but we don't need to write any logic to log out the user. In fact, I'm feeling so sure that I'm going to throw a new Exception():

  will be intercepted before getting here
Remember how "authenticators" run automatically at the beginning of every request, before the controllers? The logout process works the same way. All we need to do is tell Symfony what URL we want to use for logging out.
In security.yaml, under your firewall, add a new key: logout and, below that, path set to our logout route. So, for us, it's app_logout.
That's it! Now, whenever a user goes to the app_logout route, at the beginning of that request, Symfony will automatically log the user out and then redirect them... all before the controller is ever executed.
So... let's try it! Change the URL to /logout and... yes! The web debug toolbar reports that we are once again floating around the site anonymously.
By the way, there are a few other things that you can customize under the logout section, like where to redirect. You can find those options in the Symfony reference section.
But now, we need to talk about CSRF protection. We'll also add remember me functionality to our login form with almost no effort.